+++ Packages/KeyboardCore/Sources/KeyboardCore/KeyboardInputSettings.swift
@@ -1,0 +1,3 @@
+public enum KeyboardInputSettingsKey {
+    public static let pairedSymbolCompletionEnabled = "paired_symbol_completion_enabled"
+}